% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_release.R
\name{gen_tag}
\alias{gen_tag}
\title{Created a tag from the file version}
\usage{
gen_tag(filepath)
}
\arguments{
\item{filepath}{path to the PSD file (PSNUxIM or NAT_SUBNAT)}
}
\value{
tag name
}
\description{
Created a tag from the file version
}
\keyword{internal}
