% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_uid.R
\name{generate_uid}
\alias{generate_uid}
\title{Generate a Unique ID}
\usage{
generate_uid(codeSize = 11)
}
\arguments{
\item{codeSize}{character length for UID output (default = 11)}
}
\description{
This function is used to create a unique ID (UID) to mimic the UIDs created
and used by DATIM for the MER and other PEPFAR Structured Datasets.
}
\examples{
 generate_uid()
}
