% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minoria_mechs}
\alias{minoria_mechs}
\title{Minoria Mechanism Table}
\format{
\subsection{\code{minoria_mechs}}{

A data frame with 122 rows and 7 columns:
\describe{
\item{mech_code}{Unique implementing mechanism code}
\item{mech_name}{implementing mechanism code (from MiLB team name)}
\item{prime_partner_name}{implementing mechanism partner (from MiLB team
name)}
}
}
}
\usage{
minoria_mechs
}
\description{
This table mimics the structure of PEPFAR's MER Structured Datasets (MSD) for
mechanims, which originally originate from FACTInfo NextGen. The dummy
dataset for the (Kingdom of) Minoria has 120 (plus two dedup) mechanisms
available for use. Derived from \code{milb}.
}
\keyword{datasets}
