% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minoria_shp_ou}
\alias{minoria_shp_ou}
\title{Minoria PSNU Shape File}
\format{
\subsection{\code{minoria_shp_ou}}{

A data frame with 16 rows and 7 columns:
\describe{
\item{operatingunituid}{OU unique ID (Minoria)}
\item{operatingunit}{Operating Unit name (Minoria)}
\item{country}{Country (Minoria)}
\item{geometry}{Country level polygon shape for mapping}
}
}
}
\usage{
minoria_shp_ou
}
\description{
This dataset is the sf file for mapping (Kingdom of) Minoria at the national
level
}
\seealso{
Other shp: 
\code{\link{minoria_shp_psnu}},
\code{\link{minoria_shp_snu1}}
}
\concept{shp}
\keyword{datasets}
