% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minoria_shp_psnu}
\alias{minoria_shp_psnu}
\title{Minoria PSNU Shape File}
\format{
\subsection{\code{minoria_shp_psnu}}{

A data frame with 16 rows and 8 columns:
\describe{
\item{psnu}{Priority SNU (from MiLB city)}
\item{psnuuid}{PSNU unique ID}
\item{snu1}{Sub-National Unit 1 level below national (from MiLB league)}
\item{snu1uid}{SNU1 unique ID}
\item{operatingunit}{Operating Unit name (Minoria)}
\item{operatingunituid}{OU unique ID (Minoria)}
\item{country}{Country (Minoria)}
\item{geometry}{PSNU level polygon shape for mapping}
}
}
}
\usage{
minoria_shp_psnu
}
\description{
This dataset is the sf file for mapping (Kingdom of) Minoria by PSNU.
}
\seealso{
Other shp: 
\code{\link{minoria_shp_ou}},
\code{\link{minoria_shp_snu1}}
}
\concept{shp}
\keyword{datasets}
