% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minoria_shp_snu1}
\alias{minoria_shp_snu1}
\title{Minoria PSNU Shape File}
\format{
\subsection{\code{minoria_shp_snu1}}{

A data frame with 4 rows and 6 columns:
\describe{
\item{snu1uid}{SNU1 unique ID}
\item{snu1}{Sub-National Unit 1 level below national (from MiLB league)}
\item{operatingunit}{Operating Unit name (Minoria)}
\item{operatingunituid}{OU unique ID (Minoria)}
\item{country}{Country (Minoria)}
\item{geometry}{SNU1 level polygon shape for mapping}
}
}
}
\usage{
minoria_shp_snu1
}
\description{
This dataset is the sf file for mapping (Kingdom of) Minoria by SNU1.
}
\seealso{
Other shp: 
\code{\link{minoria_shp_ou}},
\code{\link{minoria_shp_psnu}}
}
\concept{shp}
\keyword{datasets}
