% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_create.R
\name{msk_create}
\alias{msk_create}
\title{Create the Masked Dataset}
\usage{
msk_create(filepath)
}
\arguments{
\item{filepath}{path to the MSD file}
}
\value{
dataframe with converted geography + mech info
}
\description{
Create the Masked Dataset
}
