% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_download.R
\name{msk_download}
\alias{msk_download}
\title{Download Masked Dataset}
\usage{
msk_download(folderpath, tag = "latest", launch = FALSE)
}
\arguments{
\item{folderpath}{where should the file be downloaded to?}

\item{tag}{version tag, default = "latest"}

\item{launch}{whether to launch Windows Explorer to the location after the
download completes (default = FALSE)}
}
\description{
This function download a masked dataset from GitHub
(\href{https://github.com/USAID-OHA-SI/themask}{USAID-OHA-SI/themask}) for use in
training or testing. Recommend running \code{msk_available} to see what version
is available to download.
}
\examples{
\dontrun{
#check available version
msk_available()

#download to your downloads folder
msk_download("~/Downloads)
}
\dontrun{
#download an older version
msk_download("~/Downloads, tag = "2023.06.27c")
}
}
\references{
msk_available
}
