% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_export.R
\name{msk_export}
\alias{msk_export}
\title{Export masked dataset}
\usage{
msk_export(df, filepath, output_folder)
}
\arguments{
\item{df}{df}

\item{filepath}{path to the PSD file (PSNUxIM or NAT_SUBNAT)}

\item{output_folder}{location where you want to store the new file}
}
\description{
Export masked dataset
}
\keyword{internal}
