% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_export.R
\name{msk_filename}
\alias{msk_filename}
\title{Create a file name for export}
\usage{
msk_filename(df, filepath)
}
\arguments{
\item{df}{df}

\item{filepath}{path to the PSD file (PSNUxIM or NAT_SUBNAT)}
}
\description{
Create a file name for export
}
\keyword{internal}
