% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_import.R
\name{msk_import}
\alias{msk_import}
\title{Import PSD to mask}
\usage{
msk_import(filepath)
}
\arguments{
\item{filepath}{path to a PSD file, either a PSNUxIM or NAT_SUBNAT}

\item{call}{}
}
\value{
df
}
\description{
Import PSD to mask
}
\keyword{internal}
