% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_mech.R
\name{msk_mech_map}
\alias{msk_mech_map}
\title{Create a mechanism mapping table}
\usage{
msk_mech_map(df)
}
\arguments{
\item{df}{}
}
\value{
df
}
\description{
Create a mechanism mapping table
}
\keyword{internal}
