% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_mech.R
\name{msk_mech_pull}
\alias{msk_mech_pull}
\title{Pulls from Existing Mech Mapping List on GDrive + Creates new}
\usage{
msk_mech_pull(df)
}
\arguments{
\item{df}{}
}
\value{
df
}
\description{
Pulls from Existing Mech Mapping List on GDrive + Creates new
}
\keyword{internal}
