% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_release.R
\name{msk_release}
\alias{msk_release}
\title{Upload a New Masked Dataset to GitHub}
\usage{
msk_release(filepath, output_folder)
}
\arguments{
\item{filepath}{path to the PSD file (PSNUxIM or NAT_SUBNAT) or masked file}

\item{output_folder}{location where you want to store the new file (default
stores to a temporary directory)}
}
\value{
dataframe with converted geography + mech info
}
\description{
This function is used to upload the new masked dataset to GitHub
(USAID-OHA-SI/themask) so that others can use it. This function is for
package developers use only. The data from \code{\link[=msk_created]{msk_created()}} are subset to
16 PSNUs and masked across geographic and mechanism variables. Either the
PSNUxIM or NAT_SUBNAT Structured Datasets can be masked.
}
