% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_year.R
\name{msk_year}
\alias{msk_year}
\title{Mask Fiscal Year}
\usage{
msk_year(df, n_years = 37)
}
\arguments{
\item{n_years}{number of years to add to \code{fiscal_year}, default = 37}
}
\value{
df
}
\description{
Mask Fiscal Year
}
\keyword{internal}
