% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minoria_geo}
\alias{minoria_geo}
\title{Minoria Geography Table}
\format{
\subsection{\code{minoria_geo}}{

A data frame with 16 rows and 7 columns:
\describe{
\item{operatingunit}{Operating Unit name (Minoria)}
\item{operatingunituid}{OU unique ID (Minoria)}
\item{country}{Country (Minoria)}
\item{snu1}{Sub-National Unit 1 level below national (from MiLB league)}
\item{snu1uid}{SNU1 unique ID}
\item{cop22_psnu}{Priority SNU (from MiLB city)}
\item{cop22_psnuuid}{PSNU unique ID}
}
}
}
\usage{
minoria_geo
}
\description{
This table mimics the structure of PEPFAR's MER Structured Datasets (MSD).
The dummy dataset for the (Kingdom of) Minoria has four region (\code{snu1}), each
containing four districts (\code{psnu}). Derived from \code{milb}.
}
\keyword{datasets}
