% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_download.R
\name{msk_available}
\alias{msk_available}
\title{Check the latest version available}
\usage{
msk_available()
}
\description{
This function is used to check what masked version is currently available and
will flag if there it is up to date or you should run \code{msk_create} yourself.
}
