% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_create.R
\name{msk_create}
\alias{msk_create}
\title{Create the Masked Dataset}
\usage{
msk_create(filepath, output_folder)
}
\arguments{
\item{filepath}{path to the PSD file (PSNUxIM or NAT_SUBNAT)}

\item{output_folder}{location where you want to store the new file (default =
does not export the data)}
}
\value{
dataframe with converted geography + mech info
}
\description{
This function is used to create a masked dataset for use in testing and
training. The data are subset to 16 PSNUs and masked across geographic and
mechanism variables. Either the PSNUxIM or NAT_SUBNAT Structured Datasets
can be masked.
}
\examples{
\dontrun{
#create a masked dataset
library(glamr)
library(themask)

#store path to latest MSD
path <- si_path() \%>\% return_latest("PSNU_IM")

#create a masked dataset from the PSNUxIM MSD
msk_create(path, "project1/data")
}
}
