% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_create.R
\name{msk_create}
\alias{msk_create}
\title{Create the Masked Dataset}
\usage{
msk_create(filepath, output_folder)
}
\arguments{
\item{filepath}{path to the PSD file (PSNUxIM or NAT_SUBNAT)}

\item{output_folder}{location where you want to store the new file}
}
\value{
dataframe with converted geography + mech info
}
\description{
Create the Masked Dataset
}
